Rework to avoid non-models inheriting models.
Having a regular class inherit from a Pydantic model turned out to be
more effort than it was worth, so instead we have the model be an
attribute of the model+client class, and expose attributes of the model
via accessors. This has the benefit of (pseudo-)immutability, and
making it easier to change the type of exposed attributes.